Convert dep install to a single stage script.
This required a tad more coercing to get it to behave, go env adds extra variables which I was trying to use to determine the architecture, but that doesn't work outside of go.
Add an if statement that will take care of working out the kernel arch required for dep.